requirements.txt
Django==2.1.2
xhtml2pdf --pre
djangorestframework==3.8.2
django-filter==2.0.0
djangorestframework-simplejwt==3.2.3
django-cors-headers==2.4.0